# backend/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/acs-keys/member/{memberId}": {
            "get": {
                "operationId": "ACSKeysController_findAllByMemberId",
                "summary": "Get ACS keys for specific member",
                "parameters": [
                    {
                        "name": "memberId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ACSKeyDTO"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Erroneous response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorApiResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "acs-keys"
                ],
                "security": [
                    {
                        "cookie": []
                    }
                ]
            }
        },
        "/api/acs-keys": {
            "post": {
                "operationId": "ACSKeysController_create",
                "summary": "Create ACS key",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateACSKeyDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ACSKeyDTO"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Erroneous response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorApiResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "acs-keys"
                ],
                "security": [
                    {
                        "cookie": []
                    }
                ]
            }
        },
        "/api/acs-keys/{id}": {
            "delete": {
                "operationId": "ACSKeysController_remove",
                "summary": "Delete ACS key",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful response"
                    },
                    "default": {
                        "description": "Erroneous response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorApiResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "acs-keys"
                ],
                "security": [
                    {
                        "cookie": []
                    }
                ]
            }
        },
...
    },
    "info": {
        "title": "RE: Swynca",
        "description": "B4CKSP4CE member/money management system",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "cookie": {
                "type": "apiKey",
                "in": "cookie",
                "name": "session"
            }
        },
        "schemas": {
...
            "ErrorApiResponse": {
                "type": "object",
                "properties": {
                    "statusCode": {
                        "type": "number"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "required": [
                    "statusCode",
                    "message"
                ]
            },
...
            "ACSKeyDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "memberId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "pan",
                            "uid"
                        ]
                    },
                    "key": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "memberId",
                    "type",
                    "key",
                    "name"
                ]
            },
            "CreateACSKeyDTO": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "enum": [
                            "pan",
                            "uid"
                        ]
                    },
                    "key": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "memberId": {
                        "type": "string",
                        "format": "uuid"
                    }
                },
                "required": [
                    "type",
                    "key",
                    "name",
                    "memberId"
                ]
            },
...
        }
    }
}